AppTotemClean/backend-example/requirements.txt
# Backend Django - Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Reconhecimento facial
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Banco de dados
psycopg2-binary==2.9.9

# Google Cloud
google-cloud-storage==2.10.0

# Utilitários
python-decouple==3.8
gunicorn==21.2.0 